Fix #129: Missing break in the 16 bit case, and actually make the tests test that.
Add in a warning (it's duplicate printed, but...)

+++ test/suites/sccz80/lshift.c
@@ -1,5 +1,6 @@
 #include "test.h"
 #include <stdio.h>
+#include <stdlib.h>
 
 void test_lshift16_var(void)
 {
@@ -121,27 +122,27 @@ void test_lshift32_var(void)
 
 void test_lshift16_const(void)
 {
-     long val = 1;
+     int val = 1;
 
      // RHS is folded by compiler
-     Assert( val << 0 == 1L << 0, "<<0");
-     Assert( val << 1 == 1L << 1, "<<1");
-     Assert( val << 2 == 1L << 2, "<<2");
-     Assert( val << 3 == 1L << 3, "<<3");
-     Assert( val << 4 == 1L << 4, "<<4");
-     Assert( val << 5 == 1L << 5, "<<5");
-     Assert( val << 6 == 1L << 6, "<<6");
-     Assert( val << 7 == 1L << 7, "<<7");
-     Assert( val << 8 == 1L << 8, "<<8");
-     Assert( val << 9 == 1L << 9, "<<9");
-     Assert( val << 10 == 1L << 10, "<<10");
-     Assert( val << 11 == 1L << 11, "<<11");
-     Assert( val << 12 == 1L << 12, "<<12");
-     Assert( val << 13 == 1L << 13, "<<13");
-     Assert( val << 14 == 1L << 14, "<<14");
-     Assert( val << 15 == 1L << 15, "<<15");
-     Assert( val << 16 == 1L << 16, "<<16"); // Undefined
-     Assert( val << 17 == 1L << 17, "<<17"); // Undefined
+     Assert( val << 0 == 1 << 0, "<<0");
+     Assert( val << 1 == 1 << 1, "<<1");
+     Assert( val << 2 == 1 << 2, "<<2");
+     Assert( val << 3 == 1 << 3, "<<3");
+     Assert( val << 4 == 1 << 4, "<<4");
+     Assert( val << 5 == 1 << 5, "<<5");
+     Assert( val << 6 == 1 << 6, "<<6");
+     Assert( val << 7 == 1 << 7, "<<7");
+     Assert( val << 8 == 1 << 8, "<<8");
+     Assert( val << 9 == 1 << 9, "<<9");
+     Assert( val << 10 == 1 << 10, "<<10");
+     Assert( val << 11 == 1 << 11, "<<11");
+     Assert( val << 12 == 1 << 12, "<<12");
+     Assert( val << 13 == 1 << 13, "<<13");
+     Assert( val << 14 == 1 << 14, "<<14");
+     Assert( val << 15 == 1 << 15, "<<15");
+     Assert( val << 16 == 1 << 16, "<<16"); // Undefined
+     Assert( val << 17 == 1 << 17, "<<17"); // Undefined
 }
 
 void test_lshift32_const(void)
